Created popular component
This component fetches items from spoonacular. First I check the local storage if the data's already in there it calls it to save from calling more time than needed. If not it fetch's the data then sets the poplar state to it. I can then loop through it to display certain information and images.