Added nav with search bar in it
Created navbar that sits at the top of every page. It includes a logo, basic pages, a recipe dropdown, and a search bar. The navbar changes when mobile sized to a hamburger menu to fit content better. I was able to do this by tracking the windows width, and depending once it hit the size I wanted ( 850px ) it changes the display. The search bar works similar to the recipe pages except it uses useNavigate rather than useParams.  This is because it's a search bar that's submitting results rather than with params where it's grabbing the url which for the recipe page is the items id.